+++ weblate/trans/tests/test_views.py
@@ -136,7 +136,7 @@ class ViewTestCase(RepoTestCase):
         return request
 
     def get_translation(self, language="cs"):
-        return self.component.translation_set.get(language_code=language)
+        return self.component.translation_set.get(language__code=language)
 
     def get_unit(self, source="Hello, world!\n", language="cs"):
         translation = self.get_translation(language)